+++ CHANGES.md
@@ -2,7 +2,8 @@
 Notable changes to this project.
 
 ## dev
-- added functionality to set the comment for your submission
+- added `set_comment` to set or update the comment associated with a submission
+- added `select_submission` to select a submission
 
 ## v0.1.3 - 2022-01-19
 - adjust type hints, to allow using Python versions <3.9