+++ CHANGES.md
@@ -1,7 +1,7 @@
 # Changes
 Notable changes to this project.
 
-## dev
+## v0.1.4 - 2022-02-04
 - added `set_comment` to set or update the comment associated with a submission
 - added `select_submission` to select a submission